- added experimental support for Recent Files and Sessions
git-svn-id: https://svn2.assembla.com/svn/savant/Savant/trunk@1019 096565f4-0d05-4fcd-8cbf-ff6523290f6f

+++ src/savant/view/swing/ExportImage.java
@@ -56,8 +56,8 @@ public class ExportImage {
 
         //write range at top
         RangeController range = RangeController.getInstance();
-        int start = range.getRangeStart();
-        int end = range.getRangeEnd();
+        long start = range.getRangeStart();
+        long end = range.getRangeEnd();
         String toWrite = "Range:  " + start + " - " + end;
         Graphics2D g = out.createGraphics();
         g.setColor(Color.white);
